Filter components to apply during odo push
